requirements.txt
apache-airflow==3.0.0
google-api-python-client==2.171.0
apache-airflow-providers-microsoft-azure==12.3.1